♻️ move context providers to root
