+++ apps/web/app/(entity)/[network]/layout.tsx
@@ -1,7 +1,6 @@
 // components
 import { Header } from "~/ui/header";
 import { HideBodyOverflow } from "~/ui/hide-body-overflow";
-import { RQTableProvider } from "~/ui/entity/table/table-provider";
 
 // utils
 import { getSingleNetworkCached } from "~/lib/network";
@@ -22,37 +21,35 @@ export default async function BlockLayout({
   const network = await getSingleNetworkCached(params.network);
 
   return (
-    <RQTableProvider>
-      <main
-        className="min-h-screen flex flex-col h-full"
-        style={{
-          "--color-primary": network?.config.primaryColor,
-          "--gradient-primary": network?.config.cssGradient,
-        }}
-      >
-        <HideBodyOverflow />
+    <main
+      className="min-h-screen flex flex-col h-full"
+      style={{
+        "--color-primary": network?.config.primaryColor,
+        "--gradient-primary": network?.config.cssGradient,
+      }}
+    >
+      <HideBodyOverflow />
 
-        <Header networkSlug={params.network} />
+      <Header networkSlug={params.network} />
 
-        <div className="grid lg:grid-cols-6">
-          <div
-            id="main-content"
-            tabIndex={0}
-            className={cn(
-              "mt-[65px] bg-white min-h-screen",
-              // style children but not the header nav
-              "[&>*:not(nav)]:bg-white col-span-6",
-            )}
-          >
-            {children}
-          </div>
-
-          {/* Down gradient  */}
-          <div className="fixed w-full bottom-0 z-30 left-0 h-10 !bg-gradient-to-b !rounded-none from-transparent bg-transparent to-mid-dark-100" />
-          {rightpanel}
+      <div className="grid lg:grid-cols-6">
+        <div
+          id="main-content"
+          tabIndex={0}
+          className={cn(
+            "mt-[65px] bg-white min-h-screen",
+            // style children but not the header nav
+            "[&>*:not(nav)]:bg-white col-span-6",
+          )}
+        >
+          {children}
         </div>
-      </main>
-    </RQTableProvider>
+
+        {/* Down gradient  */}
+        <div className="fixed w-full bottom-0 z-30 left-0 h-10 !bg-gradient-to-b !rounded-none from-transparent bg-transparent to-mid-dark-100" />
+        {rightpanel}
+      </div>
+    </main>
   );
 }
