--- apps/web/ui/entity/table/table-provider.tsx
@@ -1,20 +1,0 @@
-"use client";
-
-import * as React from "react";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-
-interface Props {
-  children: React.ReactNode;
-}
-
-const queryClient = new QueryClient();
-
-export function RQTableProvider({ children }: Props) {
-  return (
-    <QueryClientProvider client={queryClient}>
-      {children}
-      <ReactQueryDevtools initialIsOpen={false} position="bottom" />
-    </QueryClientProvider>
-  );
-}
